Simplify over-complicated search reducer
