+++ src/routes/Search/modules/search.js
@@ -1,5 +1,5 @@
 import * as actions from './actions'
-import reducer from './reducers'
+import reducer from './reducer'
 import query from './query'
 
 export { actions, reducer, query }
